Fix cost rate

+++ contract/views.py
@@ -2437,7 +2437,7 @@ def save_customer_service_item(request):
 
             # SRV_ACTIVE
             if (srv_active is not None):
-                field_is_modified, record = check_modified_field("CUS_SERVICE", srv_id, "SRV_ACTIVE", int(data.srv_active), int(srv_active), "E", request)
+                field_is_modified, record = check_modified_field("CUS_SERVICE", srv_id, "SRV_ACTIVE", float(data.srv_active), float(srv_active), "E", request)
                 if field_is_modified:
                     data.srv_active = srv_active
                     modified_records.append(record)
@@ -2445,7 +2445,7 @@ def save_customer_service_item(request):
 
             # SRV_RATE
             if (srv_rate is not None):
-                field_is_modified, record = check_modified_field("CUS_SERVICE", srv_id, "SRV_RATE", int(data.srv_rate), int(srv_rate), "E", request)
+                field_is_modified, record = check_modified_field("CUS_SERVICE", srv_id, "SRV_RATE", float(data.srv_rate), float(srv_rate), "E", request)
                 if field_is_modified:
                     data.srv_rate = srv_rate
                     modified_records.append(record)
@@ -2453,7 +2453,7 @@ def save_customer_service_item(request):
 
             # SRV_COST
             if (srv_cost is not None):
-                field_is_modified, record = check_modified_field("CUS_SERVICE", srv_id, "SRV_COST", int(data.srv_cost), int(srv_cost), "E", request)
+                field_is_modified, record = check_modified_field("CUS_SERVICE", srv_id, "SRV_COST", float(data.srv_cost), float(srv_cost), "E", request)
                 if field_is_modified:
                     data.srv_cost = int(srv_qty) * float(srv_rate)
                     modified_records.append(record)
